feat(blog): enhance blog styling and syntax highlighting
Updated the TailwindCSS configuration to include the typography plugin and
adjust styling for various HTML elements to improve the visual hierarchy and
readability. Added Pygments syntax highlighting to the blog post content
processing in the `load-blog-posts.py` script to enhance code readability in
blog posts. This feature provides a better user experience for readers by clearly
distinguishing code snippets from regular text.
Updated the Makefile to use the more standardized command for the TailwindCSS CLI
('tailwindcss' instead of 'tailwind').
Added styling for code blocks and inline code elements in the main CSS file to
ensure consistency with the newly introduced Pygments syntax highlighting.
Wrapped the blog post content in an HTML article tag within the
`single.jinja.html` template to improve semantic markup and accessibility.
Included the Pygments package in the `requirements.txt` file to ensure the
syntax highlighting feature is available in the production environment.
The changes were made to improve the overall aesthetic and functional quality of
the blog, provide better code readability, and ensure proper use of semantic
HTML.