refactor(config): move and restructure config classes
Moved `config.py` from a subdirectory to the project root, and
restructured the configuration loading to be object-oriented. Renamed
`Config` to `BaseConfig` as a parent class and refactored `ConfigDev` and `ConfigProd` to
inherit from it. Added `load`, `parse`, and `init_app` methods in the new `Config`
class to dynamically load and apply configurations based on the `FLASK_ENV`
environment variable. This change enables the configuration to be
accessed from both the Flask application and also from build scripts as
the Flask application can use `init_app` and scripts without an
application context can use `parse`. The `init` method in
`jackgreen_co/__init__.py` has been updated to utilize the new `init_app`
method, streamlining the app initialization process.