feat(Makefile): enhance build and deployment flexibility and improve local development configuration
Added conditional logic in the Makefile to set the BUILD_ENV variable based on
the make command goal, enhancing build and deployment flexibility. Introduced a
new MONGOSH variable for MongoDB shell commands and integrated it into the
`build-initdb` target for development environments. Updated `SERVER_NAME` in the
development configuration to "localhost:8000" for better local development
support.

These changes ensure that the correct build environment is automatically
selected based on the target being executed, making the build and deployment
process more intuitive and less error-prone. The addition of MongoDB
initialization in the development environment allows developers to work with a
fully set-up database locally, improving the development workflow. Updating the
`SERVER_NAME` to "localhost:8000" in the development configuration makes it
easier to test the application locally without changing the hosts file.