docs: add AGPLv3 license file to the repository
Include the GNU Affero General Public License (AGPL) version 3.0 to the project
to ensure that the software remains free and open source, and to comply with
licensing requirements. This license ensures that any modifications made to the
software, especially those deployed on a network server, are made available to
the community.