fix(index.jinja.html): update button links on main index page
Change the URL for the primary button to direct users to the blog posts page
instead of the main index. Change the secondary button link to direct users to
the about page instead of the contact page. These changes improve the user
experience by guiding users to the correct sections of the website.