chore(dockerfile): switch base image from python to alpine and update installation steps
Switch the base image for the Dockerfile from `python:latest` to `alpine:latest`
to reduce the image size. Update the installation steps to use `apk` for installing
Python, pip, rust, and cargo. Move the pip installation commands to consolidate into
fewer layers, and modify user and group creation commands to align with Alpine's
conventions. Add the `--chdir` option to the `gunicorn` command to set the working
directory. These changes enhance the efficiency and security of the Docker image.