refactor(nginx): update Dockerfile to use openresty
Switch to using OpenResty instead of NGINX with Lua, as it provides a more
integrated and optimized solution for running Lua scripts with NGINX. This
change simplifies the Dockerfile by reducing the number of packages installed.
Remove unnecessary module load commands from the NGINX
configuration file, which are no longer needed with OpenResty.